requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
pymupdf
requests
anthropic
python-dotenv
aiofiles
